ntfs/requirements.txt
amqp==5.2.0
celery==5.3.6
channels==4.0.0
channels-redis==4.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
daphne==4.0.0
dj-database-url==2.1.0
Django==4.1
django-celery-beat==2.5.0
django-cors-headers==3.13.0
django-cors-middleware==1.5.0
django-environ==0.9.0
django-filter==22.1
django-oauth-toolkit==2.2.0
django-rest-knox==4.2.0
django-storages==1.12.3
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-access-policy==1.3.0
drf-extra-fields==3.7.0
drf-standardized-errors==0.12.3
drf-yasg==1.21.4
flake8==6.1.0
flake8-import-order==0.18.2
flower==2.0.1
itypes==1.2.0
lxml==5.2.1
mccabe==0.7.0
msgpack==1.0.7
multidict==6.0.5
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==21.3
Pillow==10.1.0
psycopg2==2.9.5
pyflakes==3.1.0
PyJWT==2.6.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==8.0.4
pytz==2022.6
PyYAML==6.0
qrcode==7.4.2
redis==5.0.1
requests==2.28.1
requests-oauthlib==1.3.1
sentry-sdk==2.0.0
tomli==2.0.1
twilio==9.0.5
Twisted==23.8.0
whitenoise==6.5.0
django-celery-results
pika
channels_rabbitmq
msgpack
